Made the Start Menu for the Game
I made the start menu screen for the game. It contains a start button,
an instructions menu, an difficulty options menu, and a quit button. I
added two scripts to the project, just a change scene script that is
attached to the start game button and the game difficulities button and
a quit application script for the quit button.